create recordservice and move methods so batchwrite can access them
