+++ service/src/test/java/org/databiosphere/workspacedataservice/controller/SmallBatchWriteTestConfig.java
@@ -4,6 +4,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import org.databiosphere.workspacedataservice.dao.RecordDao;
 import org.databiosphere.workspacedataservice.service.BatchWriteService;
 import org.databiosphere.workspacedataservice.service.DataTypeInferer;
+import org.databiosphere.workspacedataservice.service.RecordService;
 import org.springframework.boot.test.context.TestConfiguration;
 import org.springframework.context.annotation.Bean;
 
@@ -15,7 +16,7 @@ import org.springframework.context.annotation.Bean;
 public class SmallBatchWriteTestConfig {
 
 	@Bean
-	public BatchWriteService batchWriteService(RecordDao recordDao, DataTypeInferer inf, ObjectMapper objectMapper) {
-		return new BatchWriteService(recordDao, 1, inf, objectMapper);
+	public BatchWriteService batchWriteService(RecordDao recordDao, DataTypeInferer inf, ObjectMapper objectMapper, RecordService recordService) {
+		return new BatchWriteService(recordDao, 1, inf, objectMapper, recordService);
 	}
 }
